requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
requests==2.18.4
robotframework==3.0.2
robotframework-requests==0.4.7
urllib3==1.22
